Create How_suicide_rate_is_explained
In this project, I am interested in factors that can affect the suicide rate in Poland. By using the OLS method and RStudio, I am able to come to some findings.

The model suggested that unemployment rate and the number of deaths caused by cancer can increase the rate of suicide.

The model fulfilled all the OLS assumptions. However, R-square is 0.4769 and only indicates a moderate fitness the model explaining data. Therefore, I am looking for some further analysis to improve the model